feat: module de versionning #2

+++ iodeman/src/main/java/fr/istic/possijar/Creneau.java
@@ -102,7 +102,7 @@ public class Creneau implements Serializable {
 		this.couleur = couleur;
     }
 
-	public int getCouleur() {
+	public long getCouleur() {
 		return couleur;
 	}
 }
